Progress on LAB3-payes+boosting

+++ LAB3-bayes+boosting/labfuns.py
@@ -6,7 +6,8 @@ import matplotlib.cm as cm
 from matplotlib.patches import Ellipse
 from matplotlib.colors import ColorConverter
 import random as rnd
-from sklearn.datasets.samples_generator import make_blobs
+# from sklearn.datasets.samples_generator import make_blobs
+from sklearn.datasets import make_blobs # precedent package have been replaced since
 from sklearn import decomposition, tree
 
 # import seaborn as sns
@@ -106,20 +107,20 @@ def trteSplitEven(X,y,pcSplit,seed=None):
 def fetchDataset(dataset='iris'):
     if dataset == 'iris':
         X = genfromtxt('irisX.txt', delimiter=',')
-        y = genfromtxt('irisY.txt', delimiter=',',dtype=np.int)-1
+        y = genfromtxt('irisY.txt', delimiter=',',dtype=np.int64)-1
         pcadim = 2
     elif dataset == 'wine':
         X = genfromtxt('wineX.txt', delimiter=',')
-        y = genfromtxt('wineY.txt', delimiter=',',dtype=np.int)-1
+        y = genfromtxt('wineY.txt', delimiter=',',dtype=np.int64)-1
         pcadim = 0
     elif dataset == 'olivetti':
         X = genfromtxt('olivettifacesX.txt', delimiter=',')
         X = X/255
-        y = genfromtxt('olivettifacesY.txt', delimiter=',',dtype=np.int)
+        y = genfromtxt('olivettifacesY.txt', delimiter=',',dtype=np.int64)
         pcadim = 20
     elif dataset == 'vowel':
         X = genfromtxt('vowelX.txt', delimiter=',')
-        y = genfromtxt('vowelY.txt', delimiter=',',dtype=np.int)
+        y = genfromtxt('vowelY.txt', delimiter=',',dtype=np.int64)
         pcadim = 0
     else:
         print("Please specify a dataset!")
@@ -284,7 +285,7 @@ class DecisionTreeClassifier(object):
 
     def trainClassifier(self, Xtr, yTr, W=None):
         rtn = DecisionTreeClassifier()
-        rtn.classifier = tree.DecisionTreeClassifier(max_depth=Xtr.shape[1]/2+1)
+        rtn.classifier = tree.DecisionTreeClassifier(max_depth=int(Xtr.shape[1]/2+1))
         if W is None:
             rtn.classifier.fit(Xtr, yTr)
         else:
